+++ terraqtt/config.hpp
@@ -1,0 +1,9 @@
+#ifndef TERRAQTT_CONFIG_HPP_
+#define TERRAQTT_CONFIG_HPP_
+
+// clang-format off
+#define TERRAQTT_LOG_ENABLE 1
+#define TERRAQTT_LOG_LEVEL TRACE
+// clang-format on
+
+#endif