+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [Unreleased]
+### Added
+- Better documentation
+
+### Fixed
+- Keep alive timeout
+- Bad processing of incoming data
+- Ignoring of unprocessed data
+
 ## [v0.3.0] - 2020-12-29
 ### Changed
 - Capitalize type names